Create requirements.txt
The following is a breakdown of the dependencies listed in the requirements.txt file:

torch==2.0: The PyTorch library, version 2.0, is a popular deep learning framework used for building and training neural networks.

torchvision: TorchVision is a package that provides utilities and datasets for computer vision tasks. It includes models, datasets, and transforms commonly used in vision applications.

MTCNN: MTCNN (Multi-task Cascaded Convolutional Networks) is a face detection and alignment algorithm. It is commonly used in face recognition and face analysis tasks.

numpy: NumPy is a fundamental package for scientific computing in Python. It provides support for efficient numerical operations on multi-dimensional arrays and matrices.

opencv-python: OpenCV (Open Source Computer Vision Library) is a computer vision and image processing library. The opencv-python package provides Python bindings for OpenCV, enabling you to perform various image processing tasks.

Pillow: Pillow is a powerful Python imaging library that adds support for opening, manipulating, and saving many different image file formats. It provides an easy-to-use API for image processing tasks such as resizing, cropping, and applying various filters.

matplotlib: Matplotlib is a plotting library for Python. It provides a wide variety of customizable plots, charts, and visualizations, making it useful for displaying and analyzing data.